refactored duplicate code
